Check for admin privileges
Checking for admin privileges in the layout loader does not prevent the
action from running as loaders run after the action. It must be checked
in the action itself. To avoid duplicating for every action I've added
it as a hook instead.

This does not test the actual admin role or a positive check that the
action does run for admin. To do that we would need
to create a session and it seems like the database does not get reseeded
on each test run.