Update to npm-run-all2 (#1771)
This is a community fork of npm-run-all with package maintenance and various other bug fixes applied. We're at 40k public repos adopting the fork so far, and are a designated replacement package for renovate bot. The plan to keep maintaining the project pretty much as is for the foreseeable future. I'm sending out a few PRs to notable repos that I spot to see gauge interest on switching over.